Hide the current event header when there's already a header subtitle
