+++ packages/lesswrong/components/common/Header.tsx
@@ -14,7 +14,6 @@ import classNames from 'classnames';
 import { AnalyticsContext, useTracking } from '../../lib/analyticsEvents';
 import { isEAForum, PublicInstanceSetting } from '../../lib/instanceSettings';
 import { useUnreadNotifications } from '../hooks/useUnreadNotifications';
-import { currentEventHeader } from '../../lib/publicSettings';
 
 export const forumHeaderTitleSetting = new PublicInstanceSetting<string>('forumSettings.headerTitle', "LESSWRONG", "warning")
 export const forumShortTitleSetting = new PublicInstanceSetting<string>('forumSettings.shortForumTitle', "LW", "warning")
@@ -74,19 +73,6 @@ const styles = (theme: ThemeType): JssStyles => ({
     display: 'flex',
     alignItems: 'center',
     fontWeight: isEAForum ? 400 : undefined,
-  },
-  currentEventLink: {
-    marginLeft: "1ch",
-    background: `linear-gradient(91deg,
-      ${theme.palette.text.currentEventHeader.start} 5.84%,
-      ${theme.palette.text.currentEventHeader.stop} 99.75%)
-    `,
-    backgroundClip: "text",
-    "-webkit-background-clip": "text",
-    "-webkit-text-fill-color": "transparent",
-    "&:hover": {
-      opacity: 0.8,
-    },
   },
   menuButton: {
     marginLeft: -theme.spacing.unit,
@@ -196,7 +182,6 @@ const Header = ({
   const { captureEvent } = useTracking()
   const updateCurrentUser = useUpdateCurrentUser();
   const { unreadNotifications, unreadPrivateMessages, refetch: refetchNotificationCounts } = useUnreadNotifications();
-  const currentEvent = currentEventHeader.get();
 
   const setNavigationOpen = (open: boolean) => {
     setNavigationOpenState(open);
@@ -332,14 +317,6 @@ const Header = ({
                       {hasLogo && <div className={classes.siteLogo}><Components.SiteLogo/></div>}
                       {forumHeaderTitleSetting.get()}
                     </Link>
-                    {currentEvent &&
-                      <Link
-                        to={currentEvent.link}
-                        className={classes.currentEventLink}
-                      >
-                        {currentEvent.name}
-                      </Link>
-                    }
                     <HeaderSubtitle />
                   </div>
                 </div>
