+++ packages/lesswrong/components/common/HeaderSubtitle.tsx
@@ -3,6 +3,7 @@ import { Components, registerComponent } from '../../lib/vulcan-lib';
 import { useSubscribedLocation } from '../../lib/routeUtil';
 import { Link } from '../../lib/reactRouterWrapper';
 import { isEAForum } from '../../lib/instanceSettings';
+import { currentEventHeader } from '../../lib/publicSettings';
 
 export const styles = (theme: ThemeType): JssStyles => ({
   subtitle: {
@@ -12,20 +13,34 @@ export const styles = (theme: ThemeType): JssStyles => ({
     color: theme.palette.header.text,
     borderLeft: theme.palette.border.appBarSubtitleDivider,
   },
+  currentEventLink: {
+    marginLeft: "1em",
+    background: `linear-gradient(
+      91deg,
+      ${theme.palette.text.currentEventHeader.start} 5.84%,
+      ${theme.palette.text.currentEventHeader.stop} 99.75%
+    )`,
+    backgroundClip: "text",
+    "-webkit-background-clip": "text",
+    "-webkit-text-fill-color": "transparent",
+    "&:hover": {
+      opacity: 0.8,
+    },
+  },
 });
 
 const HeaderSubtitle = ({classes}: {
   classes: ClassesType,
 }) => {
   const { currentRoute } = useSubscribedLocation();
-  if (!currentRoute) return null
+  if (!currentRoute) {
+    return null;
+  }
+
   const SubtitleComponent: any = currentRoute.subtitleComponentName ? Components[currentRoute.subtitleComponentName] : null;
   const subtitleString = currentRoute.subtitle;
   const subtitleLink = currentRoute.subtitleLink;
-  
-  if (!SubtitleComponent && !subtitleString)
-    return null;
-  
+
   if (SubtitleComponent) {
     return <SubtitleComponent isSubtitle={true} />
   } else if (subtitleLink) {
@@ -37,7 +52,17 @@ const HeaderSubtitle = ({classes}: {
       {subtitleString}
     </span>
   } else {
-    return null;
+    const currentEvent = currentEventHeader.get();
+    return currentEvent
+      ? (
+        <Link
+          to={currentEvent.link}
+          className={classes.currentEventLink}
+        >
+          {currentEvent.name}
+        </Link>
+      )
+      : null;
   }
 }
